# [Merged by Bors] - bevy_transform: Use Changed in the query for much faster transform_propagate_system