@superdump 2022-03-24
I've tested this with `transform_hierarchy -- humanoids_mixed` (modified to add back DefaultPlugins so that Tracy profiling works) which has 4000 humanoid transform hierarchies where 2000 are stationary and 2000 are actively being transformed each frame. On `main` the `transform_propagate_system` takes 8.27ms and on this PR it takes 6.68ms.